core-configuration-layer: ignore site-lisp packages on cleanup
If the user provides packages site-wide, e.g. because in NixOS they cannot be
built from elpa, but need to be built by nixpkgs, spacemacs will try to delete
these packages per default.  This commit preempts the check that
`package-delete` uses when it encounters such a package, and will just print a
message instead of throwing an error.